requirements.txt
anthropic==0.89.0
playwright==1.58.0
python-dotenv==1.0.1

# UX mode / PDF report generation only
reportlab>=4.0
